chore: move 'data_repr' and 'indent_str' to utils.py because multiple modules depend on them.
