+++ torchmeter/utils.py
@@ -3,7 +3,7 @@ import sys
 from rich import print
 from inspect import signature
 from functools import partial
-from typing import Any, Callable, Iterable, List, Optional, Tuple
+from typing import Any, Callable, Iterable, List, Optional, Sequence, Tuple, Union
 
 def perfect_savepath(origin_path:str, 
                      target_ext:Optional[str],
@@ -154,6 +154,48 @@ def dfs_task(dfs_subject:Any,
                      visited=visited)
     
     return task_res
+
+def indent_str(s:Union[str, Sequence[str]], 
+               indent:int=4, 
+               guideline:bool=True,
+               process_first:bool=True) -> str:
+    res = []
+    split_lines = s.split('\n') if isinstance(s, str) else s
+    guideline = False if len(split_lines) == 1 else guideline
+    
+    for line in split_lines:
+        indent_line = '│' if guideline else ' ' 
+        indent_line += ' '*(indent-1) + str(line)
+        res.append(indent_line)
+
+    if not process_first:
+        res[0] = res[0][indent:]
+
+    if guideline:
+        res[-1] = '└─' + res[-1][2:]
+    
+    return '\n'.join(res)
+
+def data_repr(val:Any):
+    get_type = lambda val: type(val).__name__
+
+    item_repr = lambda val_type, val: (f"[dim]Shape[/]([b green]{list(val.shape)}[/])" if hasattr(val, 'shape') else f"[b green]{val}[/]") + f" [dim]<{val_type}>[/]"
+
+    val_type = get_type(val)
+    if isinstance(val, (list, tuple, set, dict)) and len(val) > 0:
+        if isinstance(val, dict):
+            inner_repr:List[str] = [f'{item_repr(get_type(k),k)}: {data_repr(v)}' for k, v in val.items()]
+        else:
+            inner_repr:List[str] = [data_repr(i) for i in val]
+        
+        res_repr = f'[dim]{val_type}[/](' 
+        res_repr += ',\n'.join(inner_repr)
+        res_repr += ')'
+
+        return indent_str(res_repr, indent=len(f'{val_type}('), process_first=False)
+    
+    else:
+        return item_repr(val_type, val)
 
 class Verboser:
     def __init__(self,
